move AttendanceController to folder Mentor
